Support directory-based transfers
Enduro was designed around object-based storage, aka fileblobs. But one of our
two watchers is filesystem-based and lacks support for directories.

This commit updates the watchers so the include an IsDir attribute so the
processing worfklow can bundle directory-based transfers when necessary.

+++ internal/watcher/watcher.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"os"
 	"sync"
 	"time"
 
@@ -29,6 +30,9 @@ type Watcher interface {
 	RetentionPeriod() *time.Duration
 
 	StripTopLevelDir() bool
+
+	// Full path of the watched bucket when available, empty string otherwise.
+	Path() string
 
 	fmt.Stringer // It should return the name of the watcher.
 }
@@ -59,6 +63,9 @@ func (w *commonWatcherImpl) StripTopLevelDir() bool {
 type Service interface {
 	// Watchers return all known watchers.
 	Watchers() []Watcher
+
+	// Return a watcher given its name.
+	ByName(name string) (Watcher, error)
 
 	// Download blob given an event.
 	Download(ctx context.Context, w io.Writer, watcherName, key string) error
@@ -129,6 +136,10 @@ func (svc *serviceImpl) watcher(name string) (Watcher, error) {
 	return w, nil
 }
 
+func (svc *serviceImpl) ByName(name string) (Watcher, error) {
+	return svc.watcher(name)
+}
+
 func (svc *serviceImpl) Download(ctx context.Context, writer io.Writer, watcherName, key string) error {
 	w, err := svc.watcher(watcherName)
 	if err != nil {
@@ -165,6 +176,17 @@ func (svc *serviceImpl) Delete(ctx context.Context, watcherName, key string) err
 		return fmt.Errorf("error opening bucket: %w", err)
 	}
 	defer bucket.Close()
+
+	// Exceptionally, a filesystem-based watcher may be dealing with a
+	// directory instead of a regular fileblob.
+	var fi os.FileInfo
+	if bucket.As(&fi) && fi.IsDir() {
+		fw, ok := w.(*filesystemWatcher)
+		if !ok {
+			return fmt.Errorf("error removing directory: %s", err)
+		}
+		return fw.RemoveAll(key)
+	}
 
 	return bucket.Delete(ctx, key)
 }
